refactor(xds): move SDS cluster naming to translator
SDS cluster names are an xDS translation detail and have no IR
consumers. Keep the helper private to the translator and colocate its
collision and UTF-8 tests without changing generated names.
