--- internal/ir/sds.go
@@ -1,32 +1,0 @@
-// Copyright Envoy Gateway Authors
-// SPDX-License-Identifier: Apache-2.0
-// The full text of the Apache license is available in the LICENSE file at
-// the root of the repo.
-
-package ir
-
-import (
-	"crypto/sha256"
-	"encoding/hex"
-	"fmt"
-	"strings"
-	"unicode/utf8"
-)
-
-// SDSClusterNameFromURL returns the canonical xDS cluster name for an SDS URL.
-func SDSClusterNameFromURL(url string) string {
-	address := strings.TrimPrefix(url, "unix://")
-	hash := sha256.Sum256([]byte(address))
-	const maxReadablePrefixLength = 48
-
-	hashSuffix := hex.EncodeToString(hash[:16])
-	readablePrefix := strings.Trim(strings.ReplaceAll(address, "/", "_"), "_")
-	for len(readablePrefix) > maxReadablePrefixLength {
-		_, size := utf8.DecodeLastRuneInString(readablePrefix)
-		readablePrefix = readablePrefix[:len(readablePrefix)-size]
-	}
-	if readablePrefix != "" {
-		return fmt.Sprintf("sds_%s_%s", readablePrefix, hashSuffix)
-	}
-	return fmt.Sprintf("sds_%s", hashSuffix)
-}
