+++ internal/xds/translator/listener.go
@@ -877,7 +877,7 @@ func buildDownstreamQUICTransportSocket(tlsConfig *ir.TLSConfig) (*corev3.Transp
 		}
 		if cert.SDS != nil {
 			// Use external SDS server instead of ADS
-			clusterName := ir.SDSClusterNameFromURL(cert.SDS.GetURL())
+			clusterName := sdsClusterNameFromURL(cert.SDS.GetURL())
 			sdsConfig = sdsSecretConfig(cert.SDS.SecretName, clusterName)
 		}
 		tlsCtx.DownstreamTlsContext.CommonTlsContext.TlsCertificateSdsSecretConfigs = append(
@@ -920,7 +920,7 @@ func buildXdsDownstreamTLSSocket(tlsConfig *ir.TLSConfig) (*corev3.TransportSock
 		}
 		if cert.SDS != nil {
 			// Use external SDS server instead of ADS
-			clusterName := ir.SDSClusterNameFromURL(cert.SDS.GetURL())
+			clusterName := sdsClusterNameFromURL(cert.SDS.GetURL())
 			sdsConfig = sdsSecretConfig(cert.SDS.SecretName, clusterName)
 		}
 		tlsCtx.CommonTlsContext.TlsCertificateSdsSecretConfigs = append(
@@ -1004,7 +1004,7 @@ func setTLSValidationContext(tlsConfig *ir.TLSConfig, tlsCtx *tlsv3.CommonTlsCon
 
 	if tlsConfig.CACertificate.SDS != nil {
 		// Use external SDS server instead of ADS
-		clusterName := ir.SDSClusterNameFromURL(tlsConfig.CACertificate.SDS.GetURL())
+		clusterName := sdsClusterNameFromURL(tlsConfig.CACertificate.SDS.GetURL())
 		sdsConfig = sdsSecretConfig(tlsConfig.CACertificate.SDS.SecretName, clusterName)
 	}
