+++ internal/xds/translator/translator.go
@@ -1301,7 +1301,7 @@ func buildValidationContext(tlsConfig *ir.TLSUpstreamConfig) (*tlsv3.CommonTlsCo
 	if tlsConfig.CACertificate.SDS != nil {
 		// CA certificate is served by an external SDS server; use its config.
 		sds := tlsConfig.CACertificate.SDS
-		clusterName := ir.SDSClusterNameFromURL(sds.GetURL())
+		clusterName := sdsClusterNameFromURL(sds.GetURL())
 		validationContext.ValidationContextSdsSecretConfig = sdsSecretConfig(sds.SecretName, clusterName)
 	}
 	hasSANValidations := false
@@ -1396,7 +1396,7 @@ func buildXdsUpstreamTLSSocketWthCert(tlsConfig *ir.TLSUpstreamConfig, requiresA
 
 	for _, clientCert := range tlsConfig.ClientCertificates {
 		if sds := clientCert.SDS; sds != nil {
-			clusterName := ir.SDSClusterNameFromURL(sds.GetURL())
+			clusterName := sdsClusterNameFromURL(sds.GetURL())
 			sds := sdsSecretConfig(sds.SecretName, clusterName)
 			tlsCtx.CommonTlsContext.TlsCertificateSdsSecretConfigs = append(tlsCtx.CommonTlsContext.TlsCertificateSdsSecretConfigs, sds)
 			continue
